python/backend/requirements.txt
# =============================================================================
# CREDA FastAPI backend — use a DEDICATED venv (e.g. .venv), not (base) + extras.
# Mixing unrelated stacks (e.g. old Snowflake wanting cffi<2) in one venv causes resolver pain.
# Install:  .\scripts\install_backend_deps.ps1   OR   pip install -r requirements.txt
# =============================================================================

# --- Crypto stack first (PyOpenSSL 26.x + autobahn; pdfminer 20240706 allows cryptography>=36) ---
# If you need snowflake-connector-python (cffi<2), use a separate venv for CREDA only.
cryptography>=46.0.0,<47.0.0
cffi>=2.0.0,<3.0.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.2.0

# LLM / AI
langchain>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.4.0
chromadb>=1.0.0
sentence-transformers>=2.7.0

# Voice
faster-whisper>=1.1.0
edge-tts>=7.0.0
gtts>=2.5.0
pydub>=0.25.1

# Finance (casparser pins pdfminer.six — keep exact; do not add pdfplumber here)
casparser==0.8.1
pdfminer.six==20240706
pyxirr>=0.10.0
yfinance>=0.2.40
numpy>=1.26.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0

# Regulatory / PDF text — pypdf only (no pdfminer dependency; avoids pdfplumber pin wars)
pypdf>=4.0.0,<6.0.0

# Web scraping (market pulse)
httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-multipart>=0.0.9
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
